pyproject.toml
[tool.poetry]
name = "notion-to-kindle"
version = "0.1.0"
description = ""
authors = ["Sebastian Hurtado <sebassdc@gmail.com>"]
readme = "README.md"
packages = [{include = "notion_to_kindle"}]

[tool.poetry.dependencies]
python = "^3.9"
notion-client = "^2.0.0"
requests = "^2.28.2"
bs4 = "^0.0.1"
readability-lxml = "^0.8.1"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
